Refactor ACES gate module to clear SonarCloud findings
Split aptl.validation.techvault_gate into a thin public facade
(constants, GateCheck/GateReport/GateOptions, validate_scenario) plus
aptl.validation._gate_checks (the check implementations), so neither file
exceeds the 500-line limit. Bundle validate_scenario's optional knobs into a
GateOptions dataclass (9 parameters -> 4), extract the parity-manifest
complexity into small helpers (cyclomatic/cognitive complexity under the
threshold), add docstrings to every function, keep each function at three or
fewer returns, and replace Any type hints with the concrete ACES types
(Scenario, Diagnostic, BackendConformanceReport).

No behavior change: the same six checks run in the same order. Validated by
the fast gate-logic suite and the full-scenario integration gate (4 passed).
